pyproject.toml
[project]
name = "foto-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "torch>=2.10.0",
    "ultralytics>=8.4.22",
    "uvicorn>=0.41.0",
]
